Added express router and public dir for testing. #1
