+++ package.json
@@ -4,7 +4,9 @@
   "description": "A developer tool to visualize an application's component hierarchy.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node .",
+    "watch": "nodemon ."
   },
   "repository": {
     "type": "git",
@@ -18,6 +20,7 @@
   },
   "homepage": "https://github.com/jeshicawang/envision#readme",
   "dependencies": {
-    "acorn-jsx": "^3.0.1"
+    "acorn-jsx": "^3.0.1",
+    "express": "^4.14.1"
   }
 }
