+++ public/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Envision</title>
+  <link rel="stylesheet" href="default.css">
+</head>
+<body>
+  <h1>Envision</h1>
+</body>
+</html>
